requirements.txt
# Core scientific
numpy
scipy
pandas
llvmlite

# PyTorch ecosystem
pytorch-lightning
torchmetrics
torchdiffeq
torchsde

# Diffusion / generative
diffusers
k-diffusion
einops
resize-right
safetensors

# HuggingFace / NLP
transformers
tokenizers
huggingface-hub
sentence-transformers
sentencepiece
open-clip-torch

# Image / video
albumentations
imageio
opencv-python-headless
scikit-image
pillow

# ML utilities
scikit-learn
timm
accelerate

# Config
hydra-core
omegaconf
pyyaml

# Visualization
matplotlib
seaborn

# Logging
wandb
tensorboard

# Utilities
tqdm
einops